M  docker-compose.yml M  docker-freepbx/Dockerfile M  docker-freepbx/install/11-setup-config-files M  docker-freepbx/install/etc/cont-init.d/10-start M  run_freepbx.sh

+++ extra_install/asterisk_customs/docker/run_freepbx.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
+
+if [[ -z "$1" ]]; then
+	echo "Please provide the external ip (how this freepbx box is called from phones"
+	exit -1
+fi
+
+export EXTERNIP="$1"
+
 docker-compose kill freepbx
 docker-compose rm -f freepbx
 docker-compose build freepbx
 docker-compose up -d freepbx
 
 echo "Now goto http://localhost:9080 and setup username password."
+echo "Please create PJSIP extensions, not SIP"
+echo "Using hostname $EXTERNIP"
